+++ telekit/_init.py
@@ -2,6 +2,7 @@ from ._handler import Handler
 from ._chain import Chain
 from ._input_handler import InputHandler
 from ._callback_query_handler import CallbackQueryHandler
+from ._inline_buttons import InlineButton
 from ._user import User
 from .senders import BaseSender
 
@@ -15,4 +16,5 @@ def init(bot: telebot.TeleBot) -> None:
     Chain._init(bot)
     InputHandler._init(bot)
     CallbackQueryHandler._init(bot)
+    InlineButton._init(bot)
     User._init(bot)