+++ telekit/example/example_handlers/__init__.py
@@ -3,6 +3,7 @@ from . import (
     entry,
     pages,
     on_text,
+    spells,
     start,
     dsl,
     faq,
@@ -10,5 +11,5 @@ from . import (
     quiz,
     hotel,
     complete_hotel,
-    text_document,
+    text_document
 )